+++ PowerShellModulesUpdater/CHANGELOG.md
@@ -5,6 +5,21 @@
 * Changelog format follows [Keep a Changelog](https://keepachangelog.com/en).
 * Versioning adheres to [Semantic Versioning](https://semver.org/).
 
+## [1.19.0] - 2024-07-07
+
+### Added
+
+* Optionally use Dev Drive as temporary path for `Save-PSResource` with input parameter `-DevDrive`.
+  * Both for greater performance and greater resilience against antimalware file handles.
+    * Doc: <https://learn.microsoft.com/en-us/defender-endpoint/microsoft-defender-endpoint-antivirus-performance-mode>
+    * Related PSResource issue: <https://github.com/PowerShell/PSResourceGet/issues/1662>
+
+### Changes
+
+* Decrease amount of output created.
+  * Stats: Don't output stats if nothing was changed.
+  * Install modules in parallel: Only output new message if status have changed, instead of every 500ms like previously.
+
 ## [1.19.0-beta2] - 2024-04-15
 
 ### Added
@@ -69,6 +84,7 @@
 ## [1.17.0] - 2024-02-08
 
 ### Added
+
 * User context: Avoid installing to OneDrive (KFM) by overriding default install path `%USERPROFILE%\Documents\WindowsPowerShell\Modules` by:
   * Set user context environmental variable `PSModulePath` to `%LOCALAPPDATA%\Microsoft\PowerShell\Modules`.
   * Use `Microsoft.PowerShell.PSResourceGet\Save-PSResource` to override install path.
@@ -81,11 +97,13 @@
   * 240208: <https://github.com/PowerShell/PowerShell/issues/21199>
 
 ### Fixed
+
 * Function `Refresh-ModulesInstalled`: Wrongfully listed prerelease versions.
 * Use `[OutputType([System.Void])]` instead of `[OutputType($null)]` for functions that returns no output.
 * Detect, install and update scripts in non-standard directory.
 
 ### Changed
+
 * Now using the new `Microsoft.PowerShell.PSResourceGet` module instead of `PackageManagement` and `PowerShellGet`.
   * Self contained in `%LOCALAPPDATA%\Microsoft\PowerShell\PowerShellModulesUpdater`.
 * Made the script compatible with PowerShell Core (> 5.1) by:
@@ -97,29 +115,33 @@
 ## [1.16.0] - 2022-10-20
 
 ### Changed
+
 * Created a test to see if ```-AcceptLicense``` parameter is available for command ```PackageManagement\Install-Package```.
 * Converted some of the settings to input parameters (```param()```).
 
 ## [1.15.0] - 220320
 
 ### Added
+
 * Make sure not to use PowerShellGet v3 if installed.
   * It's not ready yet IMO, and it's still in beta.
 
 ### Fixed
+
 * Script would not load the newest versions of required modules ```PowerShellGet``` and ```PackageManagement```.
 * Implemented short term fix for version numbers that can't be read as ```[System.Version]```, like prereleases.
   * For instance, ```PowerShellGet v3.0.12-beta```.
   * Short term fix: Ignore such version numbers.
   * See README.md for more info.
 
-
 ## [1.14.0] - 2021-12-20
 
 ### Added
+
 * Created a ```README.md```.
 
 ### Fixed
+
 * Quick fix to handle beta/ pre-release versions
   * Like [```PowerShellGet``` pre-release ```3.0.12-beta```](https://www.powershellgallery.com/packages/PowerShellGet/3.0.12-beta).
   * In other words, version numbers that can't be parsed as ```[System.Version]```.
@@ -129,19 +151,23 @@
 ## [1.13.0] - 2021-09-14
 
 ### Added
+
 * Added installing scripts from PowerShellGallery defined in variable $ScriptsWanted.
 * Added updating already installed scripts that origin from PowerShellGallery.
 
 ### Changed
+
 * Some touchups on output to reflect that the script now does both modules and scripts.
 
 ## [1.12.0] - 2021-08-19
 
 ### Added
+
 * Started on the mission of enabling the script to install modules either in user (CurrentUser) or system (AllUsers) context.
   * System context works, user context is not properly tested yet.
 
 ### Fixed
+
 * Check for prerequired components and modules:
   * Better logic on using built in ```PowerShellGet``` and ```PackageManagement``` binaries until their modules are installed from PowerShellGallery.
 
@@ -150,12 +176,14 @@
 ## [1.11.0] - 2021-06-14
 
 ### Changed
+
 * Added quicker check for prerequired components and modules:
   * If ```PackageManagement``` exists as module type ```Script```, not just ```Binary```, we should be good.
 
 ## [1.10.0] - 2021-05-03
 
 ### Added
+
 * ~~Script version is now just ```YYMMdd``` for simplicity.~~
   * Changed back to semantic versioning later.
 * On a fresh client where the module has never run before, script can now install prerequirements and continue with the script in one run by loading newest versions of the prerequirements automatically.
@@ -166,9 +194,11 @@
   * Uninstall old version.
 
 ### Fixed
+
 * Fixed some buggy text output to terminal.
 
 ### Changed
+
 * ```Write-Information``` instead of ```Write-Output``` for outputting text to terminal.
   * ```Write-Output``` is for returning objects.
   * ```Write-Information``` has no downsides with PowerShell 6 and newer, but for PowerShell 5.1 this output stream is:
@@ -181,21 +211,25 @@
 ## [1.9.1] - 2020-10-03
 
 ### Added
+
 * Added check to see if powershellgallery.com is up and responding.
 * Added ability to harcode module versions to keep/ now remove.
 
 ## [1.9.0] - 2020-05-23
 
 ### Added
+
 * Function "Uninstall-ModuleManually" to speed up uninstall of outdated modules.
 * More useful modules to install.
 
 ### Changed
+
 * Corrected info about some of the modules in the list of modules to install.
 
 ## [1.8.0] - 2020-04-07
 
 ### Changed
+
 * Now using "PackageManagemeng\Install-Package" instead of "Install-Module", because "Install-Module" does not set error variable $? correctly if it fails.
 * Some speed ups
   * Use ".'Property'" instead of "Select-Object -ExpandProperty 'Property'
@@ -203,35 +237,42 @@
 ## [1.7.0] - 2020-03-23
 
 ### Fixed
+
 * Install-MissingSubModules
   * Lists a installed module thats not available from PowerShell Gallery anymore as missing sub module.
     * "Compare-Object" returns objects not in reference object AND difference object.
     * Use .Where instead
 
 ### Changed
+
 * Syntax fix, quotation marks on all dot properties
 * Some speed ups using .Where.
 
 ## [1.6.0] - 2019-11-28
 
 ### Added
+
 * Added setting for modules you don't want to get updated.
   * $ModulesDontUpdate.
 
 ## [1.5.2] - 2019-11-27
 
 ### Added
+
 * Added stats.
 
 ### Fixed
+
 * Minor code refactoring.
 
 ## [1.5.1] - 2019-11-14
 
 ### Added
+
 * Added "human readable" time start and time end output at the end, who reads time in the ToString('o') format anyway.
 
 ### Changed
+
 * Uses Begin, Process and End in every function.
 * Use `[OutputType]` in each function.
 * Better markdown syntax for changelog.
@@ -239,24 +280,29 @@
 ## [1.5.0] - 2019-10-22
 
 ### Added
+
 * Greatly improved speed by writing my own function for getting all installed versions of a module.
 
 ## [1.4.1] - 2019-08-08
 
 ### Added
+
 * Added option to automatically accept licenses when installing modules. Suddenly saw the first module requiring this: "Az.ApplicationMonitor".
 
 ### Fixed
+
 * Fixed code style places I saw it lagged behind. Esthetics.
 
 ## [1.4.0] - 2019-06-19
 
 ### Changed
+
 * Prerequirements are now handled automatically in the script, no need to flip a boolean for that anymore.
 
 ## [1.3.4] - 2019-06-16
 
 ### Added
+
 * Better handling of uninstalling PackageManagement if module is updated during current session.
   * If PackageManagement was updated during the same session, PowerShell must be closed and reopened before outdated version can be removed.
 * Will make sure that the user does not try to uninstall "PackageManagement" or "PowerShellGet".
@@ -264,59 +310,71 @@
 ## [1.3.3] - 2019-05-24
 
 ### Fixed
+
 * Uninstall unwanted modules did not find any modules because the script scoped variable for installed modules is not just keeping the name of the modules anymore. Fixed with a Select-Object.
 * Other bugfixes since last release, lots of small stuff I don't remember. Nothing major.
 
 ### Changed
+
 * Better logic in the update installed modules function when a parent module is updated and submodules might have been updated aswell.
 
 ## [1.3.2] - 2019-05-03
 
 ### Added
+
 * Added check for Execution Policy, will attempt to fix it if necessary by setting it to "Unrestricted" just for current process.
 
 ## [1.3.1] - 2019-04-04
 
 ### Added
+
 * Added modules
   * ImportExcel, can be used to import/modify/export to/from Excel files
 
 ### Fixed
+
 * Would fail on a clean install of Windows 10, fixed now
 
 ## [1.3.0] - 2019-04-01
 
 ### Fixed
+
 * Install-MissingSubModules
   * Used punctuation as check weather a module was a sub module. This does not work for "Microsoft.Graph.Intune" for instance. Now it removes the last punctuation and name, and checks weather this name exist in the list of installed modules.
 
 ### Changed
+
 * UpdateInstalledModules
   * If a submodule, punctuation in the name, will check once again what version is installed, in case parent module updated it.
 
 ## [1.2.0] - 2019-03-31
 
 ### Fixed
+
 * Used "$InstallMissingModules" instead of "$InstallMissingSubModules" for controlling weather the script should add missing sumodules. Type'o.
 
 ## [1.2.0] - 2019-03-24
 
 ### Added
+
 * Added three new modules
   * IntuneBackupAndRestore. John Seerden. Uses "MSGraphFunctions" module to backup and restore Intune config.
   * MSGraphFunctions. John Seerden. Wrapper for Microsoft Graph Rest API.
   * PSScriptAnalyzer. Microsoft. Used to analyze PowerShell scripts to look for common mistakes + give advice.
 
 ### Fixed
+
 * Missed a "#" on one of the output headings
 * Install missing modules: Success status after install would not display.
 
 ## [1.1.0] - 2019-03-16
 
 ### Added
+
 * Added changelog while I still remember the version history and changes.
 
 ### Fixed
+
 * Added ability to check for and install missing submodules.
   * For instance: Az currently has 81 submodules, I only had about 60 installed. Now the script will add all submodules available from PowerShellGallery not already installed on the computer.
   * Controllable by boolean $InstallMissingSubModules
@@ -324,10 +382,12 @@
 ## [1.0.1] - 2019-03-13
 
 ### Added
+
 * Added ability to install prerequirements
   * NuGet (Package Provider)
   * PowerShellGet (PowerShell Module)
   * Controllable by boolean $InstallPrerequirements
 
 ## [1.0.0] - 2019-03-10
+
 * Initial Release